Afternoon of 1/22 - testing vuforia
